Relocate response tests into new response_test.rb
