+++ test/qbwc/integration/request_generation_test.rb
@@ -6,9 +6,6 @@ class RequestGenerationTest < ActionDispatch::IntegrationTest
   def setup
     RequestGenerationTest.app = Rails.application
     QBWC.clear_jobs
-
-    $HANDLE_RESPONSE_DATA = nil
-    $HANDLE_RESPONSE_IS_PASSED_DATA = false
   end
 
   test "worker with nothing" do
@@ -45,45 +42,6 @@ class RequestGenerationTest < ActionDispatch::IntegrationTest
     assert_not_nil session.next_request
     simulate_response(session)
     assert_nil session.next_request
-  end
-
-  class HandleResponseOmitsJobWorker < QBWC::Worker
-    def requests(job)
-      {:customer_query_rq => {:full_name => 'Quincy Bob William Carlos'}}
-    end
-    def handle_response(*response)
-      $HANDLE_RESPONSE_EXECUTED = true
-    end
-  end
-
-  test "handle_response must use splat operator when omitting job argument" do
-    $HANDLE_RESPONSE_EXECUTED = false
-    QBWC.add_job(:integration_test, true, '', HandleResponseOmitsJobWorker)
-    session = QBWC::Session.new('foo', '')
-    assert_not_nil session.next_request
-    simulate_response(session)
-    assert_nil session.next_request
-    assert $HANDLE_RESPONSE_EXECUTED
-  end
-
-  class HandleResponseWithDataWorker < QBWC::Worker
-    def requests(job)
-      {:customer_query_rq => {:full_name => 'Quincy Bob William Carlos'}}
-    end
-    def handle_response(response, job, request, data)
-      $HANDLE_RESPONSE_IS_PASSED_DATA = (data == $HANDLE_RESPONSE_DATA)
-    end
-  end
-
-  test "handle_response is passed data" do
-    $HANDLE_RESPONSE_DATA = {:first => {:second => 2, :third => '3'} }
-    $HANDLE_RESPONSE_IS_PASSED_DATA = false
-    QBWC.add_job(:integration_test, true, '', HandleResponseWithDataWorker, nil, $HANDLE_RESPONSE_DATA)
-    session = QBWC::Session.new('foo', '')
-    assert_not_nil session.next_request
-    simulate_response(session)
-    assert_nil session.next_request
-    assert $HANDLE_RESPONSE_IS_PASSED_DATA
   end
 
   class MultipleRequestWorker < QBWC::Worker
