+++ test/qbwc/integration/session_test.rb
@@ -79,50 +79,4 @@ class SessionTest < ActionDispatch::IntegrationTest
     assert_equal 100, session.progress
   end
 
-  class QueryAndDeleteWorker < QBWC::Worker
-    def requests(job)
-      {:name => 'mrjoecustomer'}
-    end
-
-    def handle_response(resp, job, request, data)
-       QBWC.delete_job(job.name)
-    end
-  end
-
-  test "processes warning responses and deletes the job" do
-
-    # Add a job
-    QBWC.add_job(:query_joe_customer, true, COMPANY, QueryAndDeleteWorker)
-
-    # Simulate controller receive_response
-    ticket_string = QBWC::ActiveRecord::Session.new(QBWC_USERNAME, COMPANY).ticket
-    session = QBWC::Session.new(nil, COMPANY)
-
-    session.response = QBWC_CUSTOMER_QUERY_RESPONSE_WARN
-    assert_equal 100, session.progress
-
-    # Simulate arbitrary controller action
-    session = QBWC::ActiveRecord::Session.get(ticket_string)  # simulated get_session
-    session.save  # simulated save_session
-
-  end
-
-  test "processes error responses and deletes the job" do
-
-    # Add a job
-    QBWC.add_job(:query_joe_customer, true, COMPANY, QueryAndDeleteWorker)
-
-    # Simulate controller receive_response
-    ticket_string = QBWC::ActiveRecord::Session.new(QBWC_USERNAME, COMPANY).ticket
-    session = QBWC::Session.new(nil, COMPANY)
-
-    session.response = QBWC_CUSTOMER_QUERY_RESPONSE_ERROR
-    assert_equal 0, session.progress
-
-    # Simulate controller get_last_error
-    session = QBWC::ActiveRecord::Session.get(ticket_string)  # simulated get_session
-    session.save  # simulated save_session
-
-  end
-
 end
